adding ExtendedStates class
